refactor: lottofactory 모듈화, 접근 제어자 변경
