+++ src/main/java/lotto/model/lottofactory/LottoFactory.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lottofactory;
 
 public interface LottoFactory {
     Lotto generate();
