+++ src/test/java/lotto/model/LottosTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import lotto.model.lottofactory.AutoLottoFactory;
+
 class LottosTest {
 
     @Test
