+++ src/test/java/lotto/model/lottofactory/LottoFactoryTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lottofactory;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,6 +8,8 @@ import java.util.stream.Stream;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import lotto.model.LottoNumber;
 
 public class LottoFactoryTest {
